# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Assignment": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "value": {
            "discriminator": {
              "mapping": {
                "binary": "#/components/schemas/BinaryExpr",
                "call": "#/components/schemas/CallExpr",
                "field": "#/components/schemas/FieldExpr",
                "local": "#/components/schemas/LocalExpr",
                "number": "#/components/schemas/NumberExpr",
                "unary": "#/components/schemas/UnaryExpr"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NumberExpr"
              },
              {
                "$ref": "#/components/schemas/FieldExpr"
              },
              {
                "$ref": "#/components/schemas/LocalExpr"
              },
              {
                "$ref": "#/components/schemas/UnaryExpr"
              },
              {
                "$ref": "#/components/schemas/BinaryExpr"
              },
              {
                "$ref": "#/components/schemas/CallExpr"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "Assignment",
        "type": "object"
      },
...
      "BinaryExpr": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "binary",
            "default": "binary",
            "title": "Kind",
            "type": "string"
          },
          "left": {
            "discriminator": {
              "mapping": {
                "binary": "#/components/schemas/BinaryExpr",
                "call": "#/components/schemas/CallExpr",
                "field": "#/components/schemas/FieldExpr",
                "local": "#/components/schemas/LocalExpr",
                "number": "#/components/schemas/NumberExpr",
                "unary": "#/components/schemas/UnaryExpr"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NumberExpr"
              },
              {
                "$ref": "#/components/schemas/FieldExpr"
              },
              {
                "$ref": "#/components/schemas/LocalExpr"
              },
              {
                "$ref": "#/components/schemas/UnaryExpr"
              },
              {
                "$ref": "#/components/schemas/BinaryExpr"
              },
              {
                "$ref": "#/components/schemas/CallExpr"
              }
            ],
            "title": "Left"
          },
          "op": {
            "enum": [
              "+",
              "-",
              "*",
              "/",
              ">",
              ">=",
              "<",
              "<=",
              "=",
              "<>",
              "AND",
              "OR"
            ],
            "title": "Op",
            "type": "string"
          },
          "right": {
            "discriminator": {
              "mapping": {
                "binary": "#/components/schemas/BinaryExpr",
                "call": "#/components/schemas/CallExpr",
                "field": "#/components/schemas/FieldExpr",
                "local": "#/components/schemas/LocalExpr",
                "number": "#/components/schemas/NumberExpr",
                "unary": "#/components/schemas/UnaryExpr"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NumberExpr"
              },
              {
                "$ref": "#/components/schemas/FieldExpr"
              },
              {
                "$ref": "#/components/schemas/LocalExpr"
              },
              {
                "$ref": "#/components/schemas/UnaryExpr"
              },
              {
                "$ref": "#/components/schemas/BinaryExpr"
              },
              {
                "$ref": "#/components/schemas/CallExpr"
              }
            ],
            "title": "Right"
          }
        },
        "required": [
          "op",
          "left",
          "right"
        ],
        "title": "BinaryExpr",
        "type": "object"
      },
...
      "CallExpr": {
        "additionalProperties": false,
        "properties": {
          "args": {
            "items": {
              "discriminator": {
                "mapping": {
                  "binary": "#/components/schemas/BinaryExpr",
                  "call": "#/components/schemas/CallExpr",
                  "field": "#/components/schemas/FieldExpr",
                  "local": "#/components/schemas/LocalExpr",
                  "number": "#/components/schemas/NumberExpr",
                  "unary": "#/components/schemas/UnaryExpr"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/NumberExpr"
                },
                {
                  "$ref": "#/components/schemas/FieldExpr"
                },
                {
                  "$ref": "#/components/schemas/LocalExpr"
                },
                {
                  "$ref": "#/components/schemas/UnaryExpr"
                },
                {
                  "$ref": "#/components/schemas/BinaryExpr"
                },
                {
                  "$ref": "#/components/schemas/CallExpr"
                }
              ]
            },
            "title": "Args",
            "type": "array"
          },
          "kind": {
            "const": "call",
            "default": "call",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "args"
        ],
        "title": "CallExpr",
        "type": "object"
      },
...
      "Envelope_ScreenRunData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScreenRunData"
          },
          "serving": {
            "$ref": "#/components/schemas/ServingMeta"
          }
        },
        "required": [
          "data",
          "serving"
        ],
        "title": "Envelope[ScreenRunData]",
        "type": "object"
      },
...
      "FieldExpr": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "field",
            "default": "field",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "FieldExpr",
        "type": "object"
      },
      "FormulaAst": {
        "additionalProperties": false,
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/Assignment"
            },
            "title": "Assignments",
            "type": "array"
          },
          "output": {
            "discriminator": {
              "mapping": {
                "binary": "#/components/schemas/BinaryExpr",
                "call": "#/components/schemas/CallExpr",
                "field": "#/components/schemas/FieldExpr",
                "local": "#/components/schemas/LocalExpr",
                "number": "#/components/schemas/NumberExpr",
                "unary": "#/components/schemas/UnaryExpr"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NumberExpr"
              },
              {
                "$ref": "#/components/schemas/FieldExpr"
              },
              {
                "$ref": "#/components/schemas/LocalExpr"
              },
              {
                "$ref": "#/components/schemas/UnaryExpr"
              },
              {
                "$ref": "#/components/schemas/BinaryExpr"
              },
              {
                "$ref": "#/components/schemas/CallExpr"
              }
            ],
            "title": "Output"
          },
          "output_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Name"
          },
          "syntax_version": {
            "const": "tdx-v1",
            "default": "tdx-v1",
            "title": "Syntax Version",
            "type": "string"
          }
        },
        "required": [
          "assignments",
          "output_name",
          "output"
        ],
        "title": "FormulaAst",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalExpr": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "local",
            "default": "local",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "LocalExpr",
        "type": "object"
      },
...
      "NumberExpr": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "number",
            "default": "number",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value"
        ],
        "title": "NumberExpr",
        "type": "object"
      },
...
      "ParseIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "syntax",
              "type",
              "range",
              "future",
              "limit",
              "name"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "position": {
            "$ref": "#/components/schemas/SourcePosition"
          }
        },
        "required": [
          "code",
          "message",
          "position"
        ],
        "title": "ParseIssue",
        "type": "object"
      },
...
      "ScreenCondition": {
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "ScreenCondition",
        "type": "object"
      },
...
      "ScreenRankingCondition": {
        "additionalProperties": false,
        "properties": {
          "ascending": {
            "title": "Ascending",
            "type": "boolean"
          },
          "metric": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Metric",
            "type": "string"
          },
          "weight": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "metric",
          "ascending",
          "weight"
        ],
        "title": "ScreenRankingCondition",
        "type": "object"
      },
      "ScreenRankingPlan": {
        "additionalProperties": false,
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ScreenRankingCondition"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Conditions",
            "type": "array"
          },
          "top_n": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "conditions",
          "top_n"
        ],
        "title": "ScreenRankingPlan",
        "type": "object"
      },
      "ScreenRow": {
        "properties": {
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "pct_chg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct Chg"
          },
          "rank_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Position"
          },
          "ranking_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Score"
          },
          "ts_code": {
            "title": "Ts Code",
            "type": "string"
          }
        },
        "required": [
          "ts_code",
          "name",
          "close",
          "pct_chg"
        ],
        "title": "ScreenRow",
        "type": "object"
      },
      "ScreenRunData": {
        "properties": {
          "base_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Count"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "ranked_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranked Count"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ScreenRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScreenSourceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "ready",
              "unavailable",
              "no_date"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ScreenStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "trade_date": {
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "trade_date",
          "status",
          "base_count",
          "total",
          "steps",
          "rows",
          "next_cursor",
          "source"
        ],
        "title": "ScreenRunData",
        "type": "object"
      },
      "ScreenRunRequest": {
        "additionalProperties": false,
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ScreenCondition"
            },
            "maxItems": 26,
            "minItems": 1,
            "title": "Conditions",
            "type": "array"
          },
          "cursor": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "page_size": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "ranking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScreenRankingPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "trade_date": {
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "trade_date",
          "conditions"
        ],
        "title": "ScreenRunRequest",
        "type": "object"
      },
      "ScreenSourceInfo": {
        "properties": {
          "identity": {
            "title": "Identity",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "identity",
          "updated_at"
        ],
        "title": "ScreenSourceInfo",
        "type": "object"
      },
      "ScreenStep": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "count"
        ],
        "title": "ScreenStep",
        "type": "object"
      },
...
      "ServingMeta": {
        "properties": {
          "age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Seconds"
          },
          "built_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built At"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "state": {
            "$ref": "#/components/schemas/ServingState"
          }
        },
        "required": [
          "generation_id",
          "built_at",
          "age_seconds",
          "state",
          "message",
          "detail"
        ],
        "title": "ServingMeta",
        "type": "object"
      },
      "ServingState": {
        "enum": [
          "ready",
          "stale",
          "degraded",
          "unavailable"
        ],
        "title": "ServingState",
        "type": "string"
      },
...
      "SourcePosition": {
        "additionalProperties": false,
        "properties": {
          "column": {
            "title": "Column",
            "type": "integer"
          },
          "line": {
            "title": "Line",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "line",
          "column",
          "offset"
        ],
        "title": "SourcePosition",
        "type": "object"
      },
...
      "TdxParseData": {
        "additionalProperties": false,
        "properties": {
          "ast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FormulaAst"
              },
              {
                "type": "null"
              }
            ]
          },
          "capability": {
            "const": "parse_only",
            "default": "parse_only",
            "title": "Capability",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ParseIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "status": {
            "enum": [
              "parsed",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "syntax_version": {
            "const": "tdx-v1",
            "default": "tdx-v1",
            "title": "Syntax Version",
            "type": "string"
          },
          "translation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslationPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "unsupported": {
            "items": {
              "$ref": "#/components/schemas/UnsupportedItem"
            },
            "title": "Unsupported",
            "type": "array"
          }
        },
        "required": [
          "status",
          "ast",
          "translation",
          "issues",
          "unsupported"
        ],
        "title": "TdxParseData",
        "type": "object"
      },
      "TdxParseRequest": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          }
        },
        "title": "TdxParseRequest",
        "type": "object"
      },
      "TranslationPlan": {
        "additionalProperties": false,
        "properties": {
          "assignments": {
            "items": {
              "type": "string"
            },
            "title": "Assignments",
            "type": "array"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "functions": {
            "items": {
              "type": "string"
            },
            "title": "Functions",
            "type": "array"
          },
          "requires_full_history": {
            "title": "Requires Full History",
            "type": "boolean"
          },
          "window_lookback_bars": {
            "title": "Window Lookback Bars",
            "type": "integer"
          }
        },
        "required": [
          "fields",
          "functions",
          "assignments",
          "window_lookback_bars",
          "requires_full_history"
        ],
        "title": "TranslationPlan",
        "type": "object"
      },
      "UnaryExpr": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "unary",
            "default": "unary",
            "title": "Kind",
            "type": "string"
          },
          "op": {
            "enum": [
              "+",
              "-",
              "NOT"
            ],
            "title": "Op",
            "type": "string"
          },
          "operand": {
            "discriminator": {
              "mapping": {
                "binary": "#/components/schemas/BinaryExpr",
                "call": "#/components/schemas/CallExpr",
                "field": "#/components/schemas/FieldExpr",
                "local": "#/components/schemas/LocalExpr",
                "number": "#/components/schemas/NumberExpr",
                "unary": "#/components/schemas/UnaryExpr"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NumberExpr"
              },
              {
                "$ref": "#/components/schemas/FieldExpr"
              },
              {
                "$ref": "#/components/schemas/LocalExpr"
              },
              {
                "$ref": "#/components/schemas/UnaryExpr"
              },
              {
                "$ref": "#/components/schemas/BinaryExpr"
              },
              {
                "$ref": "#/components/schemas/CallExpr"
              }
            ],
            "title": "Operand"
          }
        },
        "required": [
          "op",
          "operand"
        ],
        "title": "UnaryExpr",
        "type": "object"
      },
      "UnsupportedItem": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "function",
              "field"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "position": {
            "$ref": "#/components/schemas/SourcePosition"
          }
        },
        "required": [
          "kind",
          "name",
          "position",
          "message"
        ],
        "title": "UnsupportedItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "rQuant Web API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/screen/run": {
      "post": {
        "operationId": "run_screen_api_v1_screen_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ScreenRunData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "运行选股条件",
        "tags": [
          "screen"
        ]
      }
    },
    "/api/v1/screen/tdx/parse": {
      "post": {
        "operationId": "parse_tdx_formula_api_v1_screen_tdx_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TdxParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TdxParseData"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "检查通达信公式",
        "tags": [
          "screen"
        ]
      }
    },
...
  }
}